renamed _v2's and _v3's to base name and mcmc_gw_extended.h to mcmc_gw.h (and cpp)
