+++ examples/injection_IMRPhenomPv2/src/injection.cpp
@@ -4,7 +4,7 @@
 #include <gwat/util.h>
 #include <gwat/detector_util.h>
 #include <gwat/io_util.h>
-#include <gwat/mcmc_gw_extended.h>
+#include <gwat/mcmc_gw.h>
 #include <gwat/mcmc_sampler.h>
 #include <gwat/standardPriorLibrary.h>
 #include <bayesship/dataUtilities.h>
@@ -275,7 +275,7 @@ int main(int argc, char *argv[])
 	double lp = log_prior->eval(&initialPosition, 0);
 	std::cout<<"Prior of initial point: "<<lp<<std::endl;
 
-	bayesship::bayesshipSampler *sampler = PTMCMC_MH_dynamic_PT_alloc_uncorrelated_GW_v2(dim, independentSamples,
+	bayesship::bayesshipSampler *sampler = PTMCMC_MH_dynamic_PT_alloc_uncorrelated_GW(dim, independentSamples,
 		ensembleSize, ensembleN, &initialPosition, (bayesship::positionInfo**)nullptr,
 		swapProb, burnIterations, burnPriorIterations, priorIterations, writePriorData, batchsize, (double **) nullptr,
 		log_prior, threads, pool, detect_number, data, psd, freq, data_lengths,
