--- include/gwat/mcmc_gw_extended.h
@@ -1,186 +1,0 @@
-#ifndef MCMC_GW_EXTENDED_H
-#define MCMC_GW_EXTENDED_H
-
-#include "util.h"
-#include "mcmc_gw.h"
-#include <bayesship/bayesshipSampler.h>
-#include <bayesship/dataUtilities.h>
-#include "adaptivelikelihoods.h"
-
-struct mcmcVariables 
-{
-	double **mcmc_noise=nullptr;
-	double *mcmc_init_pos=nullptr;
-	std::complex<double> **mcmc_data=nullptr;
-	double **mcmc_frequencies=nullptr;
-	std::string *mcmc_detectors=nullptr;
-	std::string mcmc_generation_method="";
-	int *mcmc_data_length = nullptr;
-	fftw_outline *mcmc_fftw_plans=nullptr;
-	int mcmc_num_detectors  ;
-	double mcmc_gps_time = 0;
-	double mcmc_gmst = 0;
-	int mcmc_max_dim;
-	int mcmc_min_dim;
-	MCMC_modification_struct *mcmc_mod_struct=nullptr;
-	bool mcmc_save_waveform=true;	
-	bool mcmc_intrinsic=false;	
-	int mcmc_deriv_order = 4;
-	bool mcmc_log_beta = false;
-	MCMC_user_param *user_parameters = nullptr;	
-	double maxDim;
-	bool mcmc_adaptive = false;
-	AdaptiveLikelihood *adaptivell=nullptr;
-	Quadrature *QuadMethod = NULL;
-};
-struct mcmcVariablesRJ
-{
-	double **mcmc_noise=nullptr;
-	double *mcmc_init_pos=nullptr;
-	std::complex<double> **mcmc_data=nullptr;
-	double **mcmc_frequencies=nullptr;
-	std::string *mcmc_detectors=nullptr;
-	std::string mcmc_generation_method="";
-	std::string mcmc_generation_method_extended="";
-	int *mcmc_data_length = nullptr;
-	fftw_outline *mcmc_fftw_plans=nullptr;
-	int mcmc_num_detectors  ;
-	double mcmc_gps_time = 0;
-	double mcmc_gmst = 0;
-	int mcmc_max_dim;
-	int mcmc_min_dim;
-	MCMC_modification_struct *mcmc_mod_struct=nullptr;
-	bool mcmc_save_waveform=true;	
-	bool mcmc_intrinsic=false;	
-	int mcmc_deriv_order = 4;
-	bool mcmc_log_beta = false;
-	MCMC_user_param *user_parameters = nullptr;	
-	int maxDim;
-	int minDim;
-};
-
-void PTMCMC_method_specific_prep_v2(std::string generation_method, int dimension, bool *intrinsic,MCMC_modification_struct *mod_struct);
-void RJPTMCMC_method_specific_prep_v2(std::string generation_method, int dimension, bool *intrinsic,MCMC_modification_struct *mod_struct);
-
-std::string MCMC_prep_params_v2(double *param, double *temp_params, gen_params_base<double> *gen_params, int dimension, std::string generation_method, MCMC_modification_struct *mod_struct, bool intrinsic, double gmst);
-
-
-
-
-bayesship::bayesshipSampler *  PTMCMC_MH_dynamic_PT_alloc_uncorrelated_GW_v2(
-	int dimension,
-	int independentSamples,
-	int ensembleSize,
-	int ensembleN,
-	bayesship::positionInfo *initialPosition,
-	bayesship::positionInfo **initialEnsemble,
-	double swapProb,
-	int burnIterations,
-	int burnPriorIterations,
-	int priorIterations,
-	bool writePriorData,
-	int batchSize,
-	double **priorRanges,
-	bayesship::probabilityFn *lp,
-	int numThreads,
-	bool pool,
-	int num_detectors,
-	std::complex<double> **data,
-	double **noise_psd,
-	double **frequencies,
-	int *data_length,
-	double gps_time,
-	std::string *detectors,
-	MCMC_modification_struct *mod_struct,
-	std::string generation_method,
-	std::string outputDir,
-	std::string outputFileMoniker, 
-	bool ignoreExistingCheckpoint,
-	bool restrictSwapTemperatures,
-	bool coldChainStorageOnly);
-
-bayesship::bayesshipSampler *  RJPTMCMC_MH_dynamic_PT_alloc_uncorrelated_GW_v2(
-	int minDim,
-	int maxDim,
-	int independentSamples,
-	int ensembleSize,
-	int ensembleN,
-	bayesship::positionInfo *initialPosition,
-	bayesship::positionInfo **initialEnsemble,
-	double swapProb,
-	int burnIterations,
-	int burnPriorIterations,
-	int priorIterations,
-	bool writePriorData,
-	int batchSize,
-	double **priorRanges,
-	bayesship::probabilityFn *log_prior,
-	int numThreads,
-	bool pool,
-	int num_detectors,
-	std::complex<double> **data,
-	double **noise_psd,
-	double **frequencies,
-	int *data_length,
-	double gps_time,
-	std::string *detectors,
-	MCMC_modification_struct *mod_struct,
-	std::string generation_method,
-	std::string generation_method_extended,
-	std::string outputDir,
-	std::string outputFileMoniker,
-	bool ignoreExistingCheckpoint,
-	bool restrictSwapTemperatures,
-	bool coldChainStorageOnly);
-
-double maximized_Log_Likelihood_aligned_spin_internal(std::complex<double> *data,
-				double *psd,
-				double *frequencies,
-				std::complex<double> *detector_response,
-				size_t length,
-				fftw_outline *plan
-				);
-
-double maximized_Log_Likelihood_unaligned_spin_internal(std::complex<double> *data,
-				double *psd,
-				double *frequencies,
-				std::complex<double> *hplus,
-				std::complex<double> *hcross,
-				size_t length,
-				fftw_outline *plan
-				);
-
-double Log_Likelihood_internal(std::complex<double> *data,
-			double *psd,
-			double *frequencies,
-			double *weights,
-			std::complex<double> *detector_response,
-			int length,
-			bool log10F,
-			std::string integration_method
-			);
-double Log_Likelihood_internal(std::complex<double> *data,
-	double *psd,
-	std::complex<double> *detector_response,
-	Quadrature *QuadMethod
-);
-
-
-
-double MCMC_likelihood_extrinsic(bool save_waveform, 
-	gen_params_base<double> *parameters,
-	std::string generation_method, 
-	int *data_length, 
-	double **frequencies, 
-	std::complex<double> **data, 
-	double **psd,  
-	double **weights, 
-	std::string integration_method, 
-	bool log10F,
-	std::string *detectors, 
-	int num_detectors,
-	Quadrature *QuadMethod = NULL
-	);
-
-
-#endif
